+++ src/commands/generate.rs
@@ -8,7 +8,7 @@ pub fn generate(matches: &ArgMatches<'_>) -> anyhow::Result<()> {
     let r1_dst = matches.value_of("r1-dst").unwrap();
     let r2_dst = matches.value_of("r2-dst").unwrap();
 
-    let n_records = value_t!(matches, "n-records", i32).unwrap_or_else(|e| e.exit());
+    let record_count = value_t!(matches, "record-count", i32).unwrap_or_else(|e| e.exit());
 
     info!("fq-generate start");
 
@@ -28,7 +28,7 @@ pub fn generate(matches: &ArgMatches<'_>) -> anyhow::Result<()> {
     let mut writer = PairWriter::new(w1, w2);
 
     writer
-        .write(generator, n_records)
+        .write(generator, record_count)
         .context("Could not write generated records")?;
 
     info!("fq-generate end");